+++ src/feature/search/component/index.js
@@ -1,1 +1,4 @@
 export { default as Search } from './Search/Search';
+export { default as SearchResults } from './SearchResults/SearchResults';
+export { default as SearchResultsErrorFallback } from './SearchResults/SearchResultsErrorFallback';
+export { default as SearchResultsSuspense } from './SearchResults/SearchResultsSuspense';
