+++ src/page/search/index.js
@@ -1,0 +1,1 @@
+export { default as SearchPage } from './SearchPage/SearchPage';
